Type Updates & Hello World
Changed currency amounts to be in double instead of int.

+++ src/main/java/nz/nightwind/Cauldron/Financials/Reserves/BankAccount/GenericBankAccount.java
@@ -36,27 +36,27 @@ public class GenericBankAccount extends BankAccount<GenericBankAccountNumber> {
     }
 
     @Override
-    protected boolean updateBalance(int new_balance) {
+    protected boolean updateBalance(double new_balance) {
         return super.updateBalance(new_balance);
     }
 
     @Override
-    public boolean addFunds(int funds) {
+    public boolean addFunds(double funds) {
         return super.addFunds(funds);
     }
 
     @Override
-    public boolean subtractFunds(int funds) {
+    public boolean subtractFunds(double funds) {
         return super.subtractFunds(funds);
     }
 
     @Override
-    public boolean checkAllowedBalance(int new_balance) {
+    public boolean checkAllowedBalance(double new_balance) {
         return super.checkAllowedBalance(new_balance);
     }
 
     @Override
-    public int getBalance() {
+    public double getBalance() {
         return super.getBalance();
     }
 }
